requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.2
pycodestyle==2.7.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
